UI Refactoring (Con't)
- During this process, it is advised to avoid exhaustive testing of the code as some functionality will be broken until the refactoring process has been completed.
- Do not use the banks view at this time.
